refactor: imports, packaging, unused classes, unused injections, indentation, naming, try-with-resources
