+++ back/sssscs/src/main/java/com/ib/certificate/ICertificateRepo.java
@@ -1,7 +1,5 @@
 package com.ib.certificate;
 
-import java.util.Optional;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
