+++ back/sssscs/src/main/java/com/ib/certificate/dto/CertificateRequestCreateDto.java
@@ -4,7 +4,6 @@ import java.time.LocalDateTime;
 
 import com.ib.certificate.Certificate;
 
-import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
