+++ back/sssscs/src/main/java/com/ib/certificate/request/ICertificateRequestRepo.java
@@ -1,13 +1,13 @@
-package com.ib.certificate;
+package com.ib.certificate.request;
 
 import java.util.List;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 import org.springframework.data.jpa.repository.Query;
+import org.springframework.stereotype.Repository;
 
-import com.ib.certificate.CertificateRequest.Status;
+import com.ib.certificate.request.CertificateRequest.Status;
 import com.ib.user.User;
 
 @Repository
